Move testdb sequence randomisation to shell testdb prepare
